[Update] added *support* for users running php5.2
